Add main game file with board/disc classes, images
The main game file has a basic implementation of the board and disc
with images created with Inkscape. It uses the pygame library and 
takes in the event for when the mouse is above the board and
 immediately shows an image there with the disc's center position 
following the mouse position. When the mouse is clicked and the 
disc is above the board, the disc is animated to fall down until it 
reaches the end of the board. 

The board and disc classes currently just have an attribute to store 
their representation as an image.