Create Player, PlayerHuman, PlayerComputer classes
Player is an abstract class that initializes connect4 game and
sets player name. PlayerHuman and PlayerComputer inherit from
the Player class. The get_move function in PlayerHuman class
obtains the column number and returns it. The same function
in PlayerComputer class uses minmax strategy to choose the
best move possible, though this is still in the works. In the
Board class, the get_whose_turn function returns which player
will make the next move. Also, the get_position function gets
the row in the chosen column in which the disc will drop.